"Switched from Ruby to Python and introduced load testing with Locust"
Updated the project to use a Python-based API instead of Ruby and added locust load testing which includes use of new locustfile.py for the purpose of extracting links. In web extraction functionality, different frequently visited URLs are included in the testing. Moreover, an update in 'docker-compose.yml' shows that API now depends upon 'locust' and the number of users is also specified. Also renamed 'test_results.csv' to 'final_results.csv' for clarity and understanding of test flow.