"Switch to ruby_api in Docker and improve error handling in redis connection"
The container for the API service in Docker has been switched from Python to Ruby, and the API port number has been changed accordingly both in the docker-compose file and the web environment variables.

In addition, the Ruby API connection to Redis has been improved to provide better error handling. Specifically, in case of a missing link in Redis, it is now followed up by the extraction of links and their storage in Redis, and this change in state is reflected more precisely in the logs.

Moreover, '/hello' get request was added for testing connectivity issues.